+++ tests/mock-items.ts
@@ -11,7 +11,12 @@ import {
   varchar,
 } from "drizzle-orm/pg-core";
 
-import { paradedbField, paradedbIndex } from "../src/indexing.js";
+import {
+  paradedbField,
+  paradedbIndex,
+  vector,
+  vectorField,
+} from "../src/indexing.js";
 import * as tokenizer from "../src/tokenizer.js";
 
 const int4range = customType<{ data: string; driverData: string }>({
@@ -33,6 +38,7 @@ export const mockItems = pgTable(
     lastUpdatedDate: date("last_updated_date"),
     latestAvailableTime: time("latest_available_time"),
     weightRange: int4range("weight_range"),
+    embedding: vector("embedding", { dimensions: 8 }),
   },
   (table) => [
     paradedbIndex("mock_items_search_idx").on(
@@ -49,6 +55,7 @@ export const mockItems = pgTable(
       table.lastUpdatedDate,
       table.latestAvailableTime,
       table.weightRange,
+      vectorField(table.embedding),
     ),
   ],
 );